Operator[] method made, had to change a size_t to int to get it to work in .h, and made the keys method. for both of these i of course made tests in debug for testing.

+++ HashTable.h
@@ -17,7 +17,7 @@ enum class BucketType { NORMAL, ESS, EAR };
 class HashTableBucket {
 public:
     string key;
-    size_t value;
+    int value;
     BucketType type;
     /**
     * The default constructor can simply set the bucket type to ESS.
